restrict_pages

+++ base/views.py
@@ -4,11 +4,16 @@ from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout 
 from .models import Room, Topic
 from .forms import RoomForm
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 # Create your views here.
 
 
 def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+    
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -28,7 +33,9 @@ def loginPage(request):
     return render(request, 'base/login_register.html', context)
 
 
-
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -51,6 +58,7 @@ def room(request, pk):
     return render(request, 'base/room.html', context)
 
 
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -63,11 +71,14 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
-
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id = pk)
     form = RoomForm(instance = room )
     
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!!!")
+          
     if request.method == 'POST':
         form = RoomForm(request.POST, instance = room)
         if form.is_valid():
@@ -78,9 +89,13 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
-
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id = pk)
+    
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!!!")
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
